backend/requirements.txt
# FastAPI Backend Dependencies

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai>=1.12.0
python-multipart==0.0.6
pydantic>=2.5.3
httpx>=0.24.0
PyPDF2==3.0.1
Pillow>=10.2.0
pytesseract>=0.3.10
pdf2image>=1.16.3

# Database (PostgreSQL + SQLAlchemy)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
